+++ CHANGELOG.md
@@ -6,6 +6,9 @@ GAMSPy 1.5.1
 - General
   - Fix the bugs in dynamic set assignments.
   - Allow loadRecordsFromGdx to domain forward.
+  - Enhance bound propagation for `MaxPool2d` and `MinPool2d` classes.
+- Testing
+  - Add bound propagation tests for `MaxPool2d` and `MinPool2d` classes.
 - Documentation
   - Update embedding Neural Network documentation.